split out googlefonts conditions into a separate file
